move session restore to init module
